feat(application): centralize troubleshooting session safety
