+++ internal/opcua/write.go
@@ -7,6 +7,18 @@ import (
 	"github.com/gopcua/opcua/ua"
 )
 
+// WriteRejectedError reports a server response that conclusively rejected a
+// Variable Node Write. Unlike a transport error, it proves the mutation did
+// not complete.
+type WriteRejectedError struct {
+	NodeID string
+	Status string
+}
+
+func (e *WriteRejectedError) Error() string {
+	return fmt.Sprintf("write %s rejected: %s", e.NodeID, e.Status)
+}
+
 func (c *gopcuaClient) WriteValue(ctx context.Context, nodeID string, value ScalarValue) error {
 	if c.client == nil {
 		return ua.StatusBadServerNotConnected
@@ -31,7 +43,7 @@ func (c *gopcuaClient) WriteValue(ctx context.Context, nodeID string, value Scal
 		return fmt.Errorf("write %s failed: no result", nodeID)
 	}
 	if response.Results[0] != ua.StatusOK {
-		return fmt.Errorf("write %s failed: %s", nodeID, response.Results[0])
+		return &WriteRejectedError{NodeID: nodeID, Status: fmt.Sprint(response.Results[0])}
 	}
 	return nil
 }
